Basic Factory Design pattern:
factory design pattern uses super class that has various sub classes and depending on input returns on specific sub class.
advantage of this pattern is sub classes does not have to be instantiated on client code. and changes can be made directly to the super class without requirement for refactoring instantiated codes.